Copied some emulated motion plus stuff over from the old wiimote plugin. Not at all functional currently.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@6660 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/HW/WiimoteEmu/Attachment/Nunchuk.h
@@ -30,7 +30,7 @@ private:
 	Buttons*		m_buttons;
 	AnalogStick*	m_stick;
 
-	unsigned int	m_shake_step[3];
+	u8	m_shake_step[3];
 	
 	UDPWrapper* const m_udpWrap;
 };
